+++ src/e2e/app/cli/repo-tests/src/test_smart_transfer_protocol.rs
@@ -951,7 +951,7 @@ async fn test_smart_push_all_smart_pull_all(
               REQUIRED INT64 offset;
               REQUIRED INT32 op;
               REQUIRED INT64 system_time (TIMESTAMP(MILLIS,true));
-              OPTIONAL INT64 match_time (TIMESTAMP(MILLIS,true));
+              REQUIRED INT64 match_time (TIMESTAMP(MILLIS,true));
               OPTIONAL INT64 match_id;
               OPTIONAL BYTE_ARRAY player_id (STRING);
               OPTIONAL INT64 score;
@@ -974,7 +974,7 @@ async fn test_smart_push_all_smart_pull_all(
               OPTIONAL INT64 offset;
               REQUIRED INT32 op;
               REQUIRED INT64 system_time (TIMESTAMP(MILLIS,true));
-              OPTIONAL INT64 match_time (TIMESTAMP(MILLIS,true));
+              REQUIRED INT64 match_time (TIMESTAMP(MILLIS,true));
               OPTIONAL INT64 place (INTEGER(64,false));
               OPTIONAL INT64 match_id;
               OPTIONAL BYTE_ARRAY player_id (STRING);
@@ -1249,7 +1249,7 @@ async fn test_smart_push_recursive_smart_pull_recursive(
               REQUIRED INT64 offset;
               REQUIRED INT32 op;
               REQUIRED INT64 system_time (TIMESTAMP(MILLIS,true));
-              OPTIONAL INT64 match_time (TIMESTAMP(MILLIS,true));
+              REQUIRED INT64 match_time (TIMESTAMP(MILLIS,true));
               OPTIONAL INT64 match_id;
               OPTIONAL BYTE_ARRAY player_id (STRING);
               OPTIONAL INT64 score;
@@ -1272,7 +1272,7 @@ async fn test_smart_push_recursive_smart_pull_recursive(
               OPTIONAL INT64 offset;
               REQUIRED INT32 op;
               REQUIRED INT64 system_time (TIMESTAMP(MILLIS,true));
-              OPTIONAL INT64 match_time (TIMESTAMP(MILLIS,true));
+              REQUIRED INT64 match_time (TIMESTAMP(MILLIS,true));
               OPTIONAL INT64 place (INTEGER(64,false));
               OPTIONAL INT64 match_id;
               OPTIONAL BYTE_ARRAY player_id (STRING);
@@ -1529,7 +1529,7 @@ async fn test_simple_push_to_s3_smart_pull(
               REQUIRED INT64 offset;
               REQUIRED INT32 op;
               REQUIRED INT64 system_time (TIMESTAMP(MILLIS,true));
-              OPTIONAL INT64 match_time (TIMESTAMP(MILLIS,true));
+              REQUIRED INT64 match_time (TIMESTAMP(MILLIS,true));
               OPTIONAL INT64 match_id;
               OPTIONAL BYTE_ARRAY player_id (STRING);
               OPTIONAL INT64 score;